Consolidating namespaces.

+++ Source/HoloHands/Native/Rendering/Shaders/Quad.ps.hlsl
@@ -12,8 +12,8 @@ min16float4 main(PixelShaderInput input) : SV_TARGET
    const float min = 0.00;
    const float max = 0.020;
 
-   min16float depth = quadTexture.Sample(quadSampler, input.texCoord).r;
-   min16float color = (depth - min) / (max - min); //normalize to min, max rang.e
+   float depth = quadTexture.Sample(quadSampler, input.texCoord).r;
+   float color = (depth - min) / (max - min); //normalize to min, max rang.e
    color = 1 - color; //invert.
 
    if (color <= 0 || color >= 1.0)
